Rework plugin loading
The plugins to load are semi-configurable.  Actually, they're not, but
its obvious how they could be.

Plugins are loaded in a defined order, meaning events are processed in a
defined order.

Loaded plugins are automatically registered.
